requirements.txt
langchain
google-generativeai 
langchain-google-genai
selenium
unstructured
streamlit
python-dotenv